% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonAbsValues.R
\name{jsonAbsValues}
\alias{jsonAbsValues}
\title{receives a SPSS file with the evaluation data and writes a file with the absolute values for each course}
\usage{
jsonAbsValues(spss_file, folder)
}
\arguments{
\item{spss_file}{the absolute path to a SPSS file with .sav extension}

\item{folder}{a path to an existing folder, where the generated files must be saved}
}
\value{
None
}
\description{
receives a SPSS file with the evaluation data and writes a file with the absolute values for each course
}
\examples{
   \dontrun{
   Calling jsonAbsValues on Windows operational system:
   jsonAbsValues("C:\\\\Folder\\\\WS1718.sav", "C:\\\\Folder\\\\")}
}
