% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonMeanValues.R
\name{jsonMeanValues}
\alias{jsonMeanValues}
\title{receives a SPSS file with the evaluation data and writes a file with the mean values for each variable of each course}
\usage{
jsonMeanValues(spss_file, path_to_json)
}
\arguments{
\item{spss_file}{the absolute path to a SPSS file with .sav extension}

\item{path_to_json}{a path with the name to be given to the new file ending with the .json extension}
}
\value{
None
}
\description{
receives a SPSS file with the evaluation data and writes a file with the mean values for each variable of each course
}
\examples{
   \dontrun{
   Calling jsonMeanValues on Windows operational system:
   jsonMeanValues("C:\\\\Folder\\\\WS1718.sav", "C:\\\\Folder\\\\eval.json")}

}
